Use of B_TRANSLATE_COMMENT (#42)
It's otherwise hard to translate if you don't know/see
the GUI of Slayer.

Using a pop-up menu or spinner widget restricts the freedom
of translation, as the widget always has to be at the end.
In German, e.g. it'd otherwise be:
   "Fenster in Arbeitsfläche '3' öffnen."
Adding a ":" ensures the widget is after it, at the end of
the line.